Made the OCI8Statement class final
